Add files via upload
Let us study the data and identify some conditions of Loan Accepting and Loan Rejecting
When a person applies for a loan, Either the bank accepts the loan application, or it gets rejected. In the data if we see the loan status variable, we can clearly see that-
Loan Accepted has also 3 criteria:
• Fully paid: Applicant has fully paid the loan (the principal and the interest rate)
• Current: Applicant is in the process of paying the installments, i.e. the tenure of the loan is not yet completed. These candidates are not labelled as 'defaulted'.
• Charged-off: Applicant has not paid the installments in due time for a long period of time, i.e. he/she has defaulted on the loan
Loan rejected: The company had rejected the loan (because the candidate does not meet their requirements etc.). Since the loan was rejected, there is no transnational history of those applicants with the company and so this data is not available with the company (and thus in this dataset).

using all the sheets that we analysed, I created a visual dashboard showing every analysis on a single pane.
The dashboard is set to be filtered according to be Geographical Distribution of loan amount and it will show all the other graphs updated for the selected location.
As soon as u leave the selection, it comes to the original state.
Attached is the screenshot for the Dashboard.
Conclusion:
In other words, the company wants to understand the driving factors (or driver variables) behind loan default, i.e. the variables which are strong indicators of default. The company can utilise this knowledge for its portfolio and risk assessment.